Gewünschte Radius von User wird nun bei Abfrage der Partys genommen. Anpassen des Models UserEinstellungen (double zu float). Anpassen des Models Party (hinzufügen von HostId).

+++ App2Night/ModelsEnums/Model/UserEinstellungen.cs
@@ -8,7 +8,7 @@ namespace App2Night.ModelsEnums.Model
 {
     class UserEinstellungen
     {
-        public double Radius { get; set; }
+        public float Radius { get; set; }
 
     }
 }
